+++ src/SourceBuild/content/test/Microsoft.DotNet.SourceBuild.SmokeTests/Microsoft.DotNet.SourceBuild.SmokeTests.csproj
@@ -13,7 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include=$(TasksDir)Microsoft.DotNet.UnifiedBuild.Tasks\Utilities.cs" />
+    <Compile Include="$(TasksDir)Microsoft.DotNet.UnifiedBuild.Tasks\PdbUtilities.cs" />
   </ItemGroup>
 
   <ItemGroup>
